+++ src/decorations.ts
@@ -1,86 +1,41 @@
-import { window, OverviewRulerLane, DecorationRangeBehavior } from 'vscode';
+import {
+  window,
+  OverviewRulerLane,
+  DecorationRangeBehavior,
+  ExtensionContext,
+  TextEditorDecorationType,
+} from 'vscode';
 
-export function failingItName() {
+function createTestStateDecoration(
+  context: ExtensionContext,
+  icon: string,
+  overviewRulerColor: string
+): TextEditorDecorationType {
   return window.createTextEditorDecorationType({
-    overviewRulerColor: 'red',
+    overviewRulerColor,
+    gutterIconPath: context.asAbsolutePath(`./src/icons/${icon}.svg`),
     overviewRulerLane: OverviewRulerLane.Left,
-    light: {
-      before: {
-        color: '#FF564B',
-        contentText: '✘ ',
-      },
-    },
-    dark: {
-      before: {
-        color: '#AD322D',
-        contentText: '✘ ',
-      },
-    },
     rangeBehavior: DecorationRangeBehavior.ClosedClosed,
   });
 }
 
-export function skipItName() {
-  return window.createTextEditorDecorationType({
-    overviewRulerColor: 'yellow',
-    overviewRulerLane: OverviewRulerLane.Left,
-    light: {
-      before: {
-        color: '#fed37f',
-        contentText: '○ ',
-      },
-    },
-    dark: {
-      before: {
-        color: '#fed37f',
-        contentText: '○ ',
-      },
-    },
-    rangeBehavior: DecorationRangeBehavior.ClosedClosed,
-  });
+export function passingItName(context: ExtensionContext): TextEditorDecorationType {
+  return createTestStateDecoration(context, 'passing', 'green');
 }
 
-export function passingItName() {
-  return window.createTextEditorDecorationType({
-    overviewRulerColor: 'green',
-    overviewRulerLane: OverviewRulerLane.Left,
-    light: {
-      before: {
-        color: '#3BB26B',
-        contentText: '✔ ',
-      },
-    },
-    dark: {
-      before: {
-        color: '#2F8F51',
-        contentText: '✔ ',
-      },
-    },
-    rangeBehavior: DecorationRangeBehavior.ClosedClosed,
-  });
+export function failingItName(context: ExtensionContext): TextEditorDecorationType {
+  return createTestStateDecoration(context, 'failing', 'red');
 }
 
-export function notRanItName() {
-  return window.createTextEditorDecorationType({
-    overviewRulerColor: 'darkgrey',
-    overviewRulerLane: OverviewRulerLane.Left,
-    dark: {
-      before: {
-        color: '#9C9C9C',
-        contentText: '○ ',
-      },
-    },
-    light: {
-      before: {
-        color: '#7C7C7C',
-        contentText: '○ ',
-      },
-    },
-    rangeBehavior: DecorationRangeBehavior.ClosedClosed,
-  });
+export function skipItName(context: ExtensionContext): TextEditorDecorationType {
+  return createTestStateDecoration(context, 'skip', 'yellow');
 }
 
-export function failingAssertionStyle(text: string) {
+export function notRanItName(context: ExtensionContext): TextEditorDecorationType {
+  return createTestStateDecoration(context, 'unknown', 'darkgrey');
+}
+
+export function failingAssertionStyle(text: string): TextEditorDecorationType {
   return window.createTextEditorDecorationType({
     isWholeLine: true,
     overviewRulerColor: 'red',